fix(Lib): move Supyrb library to Editor folder
The Supyrb library adds extensions to a type that is only available
in the Editor, not in builds. To fix compilation issues while
creating builds the library has been moved into an Editor folder
which ensures the extensions are only compiled and available in the
Editor, too.
