+++ Scripts/Lib/Supyrb/Editor.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c44ff70140148a14abe4d1da5ac48b97
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
